deleted Gamemanger
