+++ src/test/java/at/aau/se2/cluedo/models/GameManagerTest.java
@@ -1,7 +1,7 @@
 package at.aau.se2.cluedo.models;
 
 import at.aau.se2.cluedo.models.cards.BasicCard;
-import at.aau.se2.cluedo.models.gamemanager.CluedoGame;
+import at.aau.se2.cluedo.models.gamemanager.GameManager;
 import at.aau.se2.cluedo.models.gameobjects.Player;
 import at.aau.se2.cluedo.models.gameobjects.SecretFile;
 import org.junit.jupiter.api.BeforeEach;
@@ -14,11 +14,11 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class GameManagerTest {
 
-    private CluedoGame gameManager;
+    private GameManager gameManager;
 
     @BeforeEach
     void setUp() {
-        gameManager = new CluedoGame("6");
+        gameManager = new GameManager("6");
         gameManager.initilizeGame();
 
         ArrayList<BasicCard> rooms = new ArrayList<>();
